update some style and ==> User auth by name and pass

+++ server/routes/api/todos-routes.js
@@ -5,13 +5,23 @@ const Todo = require("../../models/todoSchema");
 
 // GET
 router.get("/", (req, res) => {
-  Todo.find()
-    .then((response) => {
-      res.send(response);
-    })
-    .catch((err) => {
-      console.log(err);
-    });
+  if (req.isAuthenticated()) {
+    Todo.find()
+      .then((response) => {
+        res.json({data: response, status: "user"});
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  } else {
+    Todo.find()
+      .then((response) => {
+        res.json({data: response, status: "notuser"});
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  }
 });
 
 // POST
@@ -29,19 +39,27 @@ router.post("/", (req, res) => {
 
 // DELETE
 router.delete("/:id", (req, res) => {
-  Todo.findByIdAndDelete(req.params.id)
-    .then((response) => {
-      // console.log(response);
-      res.status(200).send();
-    })
-    .catch((err) => console.log(err));
+  if (req.isAuthenticated()) {
+    Todo.findByIdAndDelete(req.params.id)
+      .then((response) => {
+        // console.log(response);
+        res.status(200).send();
+      })
+      .catch((err) => console.log(err));
+  } else {
+    res.send("You are not authonticated!");
+  }
 });
 
 // Get specific
 router.get("/:id", (req, res) => {
-  Todo.findById(req.params.id)
-    .then((response) => res.send(response))
-    .catch((err) => res.send(err));
+  if (req.isAuthenticated()) {
+    Todo.findById(req.params.id)
+      .then((response) => res.send(response))
+      .catch((err) => res.send(err));
+  } else {
+    res.send("You are not authenticated!");
+  }
 });
 // UPDATE or upsert
 // upsert (if not find id it will create new todo)
